Iteration 4
Setting up the rest of the game essentially. Added in rings (the first object with full animations) and an enemy which will cause Sonic's rings to reduce to zero if he is not jumping. If he does jump, it should destroy the enemy. To do still, make rings fly out for recollection, figure out how to get Sonic to fly back so that way he's not awkwardly in the enemy, and overall add in the rest of the project.